refactor utils and add suspend
