+++ src/mcp_agent/cli/cloud/commands/servers/list/main.py
@@ -8,7 +8,7 @@ from rich.panel import Panel
 
 from mcp_agent.cli.core.utils import run_async
 from mcp_agent.cli.mcp_app.api_client import MCPApp, MCPAppConfiguration
-from ..utils import (
+from ...utils import (
     setup_authenticated_client,
     validate_output_format, 
     handle_server_api_errors,
